+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     
 ### Changes 
 
+    - The record flow selector combo is now next to the tabs
     - The threads selector is now a menu button instead of a list view to gain some UI space
     
 ### Bugs fixed